The slowest expressions CI measured, split before a newer Xcode refuses one
Run 229 built with -warn-long-expression-type-checking=300 and listed
the expressions over 300 ms. The owner's Xcode had already failed on one
CI accepted, so the slowest are split into typed steps now: the brazier's
flicker (2.1 s on CI: a sum of two sines on a Float inside a CGFloat
init), the arena inlay's spoke angles (1.1 s), the Hall of Ka's feed
words (two nested ternaries of interpolated strings, and the delay's sum
inside asyncAfter, 0.95 s), the battle's MVP pick (0.59 s), the decor
sheet's pill colours (0.41 s), the cloth step's spring sum (0.41 s) and
the HUD's two sizes (0.32 s each). Nothing computed changes.

+++ Pantheon/Render/BattleSceneController.swift
@@ -1371,8 +1371,8 @@ final class BattleSceneController: NSObject {
     /// raid's barrier row 24 — so the bar ends 62 points down at most and
     /// the chips under it 98. Change one there, change it here.
     private static let hudPadding: CGFloat = 8
-    private static let hudControls = CGSize(width: 3 * 36 + 2 * 6, height: 36)
-    private static let hudSkills = CGSize(width: 3 * 60 + 2 * 10, height: 64)
+    private static let hudControls = CGSize(width: CGFloat(3 * 36 + 2 * 6), height: 36)
+    private static let hudSkills = CGSize(width: CGFloat(3 * 60 + 2 * 10), height: 64)
     private static let hudChipsLength: CGFloat = 250
     /// How far a float keeps from the HUD.
     private static let hudClearance: CGFloat = 6
